Revert "chelper: Move cartesian and delta kinematics code to their own C files"
This reverts commit c74e232739f9c4757800dc26243daf772809c814.
